refactoring. Изменяет структуру папок и убирает тип.
