+++ packages/client/src/actions/forum.ts
@@ -1,16 +1,12 @@
-import { AppDispatch, AppThunk } from 'src/store';
+import { AppDispatch } from 'src/store';
 
-import { ForumApiService } from 'src/services/api';
-import { setForum } from 'src/services/reducers/forum/forum-slice';
+import { forumApi } from 'src/api';
+import { setForum } from 'src/services/reducers/forum-slice';
 
-const api = new ForumApiService();
-// TODO нужно пофиксить тему с типом
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-export const fetchForumPage: AppThunk =
+export const fetchForumPage =
   (page: number) => async (dispatch: AppDispatch) => {
     try {
-      const threads = await api.getForum(page);
+      const threads = await forumApi.getForum(page);
       dispatch(
         setForum({
           page,
