+++ packages/client/src/api/forum.ts
@@ -1,7 +1,7 @@
 import { threads } from 'src/mock/forum-threads';
 import { ForumItemProps } from 'components/forum-item/forum-item';
 
-export class ForumApiService {
+class ForumApi {
   getForum = (
     pageNumber: string | number = 1
   ): Promise<Array<ForumItemProps>> => {
@@ -10,3 +10,5 @@ export class ForumApiService {
     });
   };
 }
+
+export const forumApi = new ForumApi();
