+++ packages/client/src/api/index.ts
@@ -1,2 +1,3 @@
 export { authApi } from './auth';
 export { userApi } from './user';
+export { forumApi } from './forum';
