--- packages/client/src/services/api/index.ts
@@ -1,1 +1,0 @@
-export { ForumApiService } from './forum/forum-api-service';
