+++ packages/client/src/services/reducers/index.ts
@@ -1,5 +1,5 @@
 import gameReducer from './game/game-slice';
-import forumReducer from './forum/forum-slice';
+import forumReducer from './forum-slice';
 
 const reducer = {
   game: gameReducer,
